fix: 4-1 inputMoving 함수 이동(App->Inputwiew)
